Control StatusUI contents by common object class

+++ Assets/Scripts/Model/Character/Player/DispStatus.cs
@@ -3,14 +3,14 @@ public struct DispStatus
     public int level;
     public float exp;
     public float expToNextLevel;
-    public int attack;
-    public int equipR;
-    public int equipL;
-    public int armor;
-    public int shield;
-    public int magic;
-    public int resistFire;
-    public int resistIce;
-    public int resistDark;
-    public int resistLight;
+    public float attack;
+    public float equipR;
+    public float equipL;
+    public float armor;
+    public float shield;
+    public float magic;
+    public float resistFire;
+    public float resistIce;
+    public float resistDark;
+    public float resistLight;
 }
